blockstore: Split column traits & types out
